[SYCL][XPTI] Unify data types used for code location
Column and line numbers used to be represented as different data types
(including both size and sign) in different parts of code.

This commit aligned them all to `uint32_t` which seems to be used under
the hood by both XPTI and UR.

The only exception here is `code_location` struct fields: since it is a
part of `handler`, we can't change data types there just yet without it
breaking ABI. However, all APIs had been changed to use the new unified
types to reduce amount of type casts necessary in the code that uses
them.

This PR belongs to a series which prepares SYCL RT codebase to be
buildable with `-Wconversion` flag that is required per our SDL
processes.

+++ sycl/source/detail/scheduler/commands.cpp
@@ -1022,9 +1022,9 @@ void Command::copySubmissionCodeLocation() {
   if (TData.functionName())
     MSubmissionFunctionName = TData.functionName();
   if (MSubmissionFileName.size() || MSubmissionFunctionName.size())
-    MSubmissionCodeLocation = {
-        MSubmissionFileName.c_str(), MSubmissionFunctionName.c_str(),
-        (int)TData.lineNumber(), (int)TData.columnNumber()};
+    MSubmissionCodeLocation = {MSubmissionFileName.c_str(),
+                               MSubmissionFunctionName.c_str(),
+                               TData.lineNumber(), TData.columnNumber()};
 #endif
 }
